refactor: point at the relocated salary-formatter import path
apps/desktop/src/renderer/lib/salary.ts moved to lib/format-salary.ts in
7051055b (folded into that commit's staged changes from a shared worktree) so
a second feature (BestMatchRow) can reuse it without an illegal
features/applications -> features/best-matches cross-feature import. Update
the two references that still pointed at the old path.
